Move test from node_test.go -> assigner_test.go
